+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Calendar Versioning](https://calver.org).<br>
 - CentOS Stream 8 target
 - `--parallel-limit` limit of plans to be run in parallel, default 42
 - query for task ID, warn and exit if build doesn't match requested test artifact
+- make builds available in the copr.fedorainfracloud.org repositories tesar and tesar-devel respectively
+
 
 ### Changed
 - `-c/--compose` to `-t/--target`